refactor(app): extract review approval controller (#346)
Extract ApproveReview.vue's state/load/modal/submission/reset/discard
orchestration into useApproveReviewController.ts and the
synopsis/comment/category/problematic dirty-state comparisons into
reviewApprovalSnapshots.ts, bringing the view from 841 to 419 lines.
Behavior-preserving: every defineExpose name, useReviewApprovalActions,
useReviewHelpers, and the lazy $refs/axios injection bridge are unchanged.
Adds order-insensitive snapshot-comparison tests plus dirty-modal
preventDefault/confirm and status update-vs-create coverage.
